+++ src/app/integrations/storage/index.ts
@@ -5,8 +5,9 @@ export {
   writeStoragePreference
 } from './preferences'
 export type { StoragePreferences } from './preferences'
-export { S3_STORAGE_PROVIDER, storageProviderRegistry } from './providers'
+export { BUNNY_STORAGE_PROVIDER, S3_STORAGE_PROVIDER, storageProviderRegistry } from './providers'
 export { defineStorageProvider, StorageProviderRegistry } from './registry'
+export { createBunnyStorageAdapter } from './bunny/adapter'
 export { createS3StorageAdapter } from './s3/adapter'
 export { storageThumbnailMimeType } from './thumbnail'
 export type { S3StorageAdapter } from './s3/adapter'